feat: modificacao do EventoController e EventoService e criacao de DTOs de detalhar

+++ src/main/java/Projeto/Gerenciador_Eventos/repository/CheckIn.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CheckIn extends JpaRepository<CheckIn, Integer>{
+public interface CheckIn extends JpaRepository<CheckIn, Long>{
 	
 	List<CheckIn> findAllByIdStatusCheckIn(Integer idStatusCheckIn);
 }
